# src/handlers/oapi_handler.go
package handlers

import (
	"html/template"
	"net/http"
	"strconv"
	"strings"

	"monolith-chi-htmx-sqlite/src/gen"
	"monolith-chi-htmx-sqlite/src/middleware"
	"monolith-chi-htmx-sqlite/src/models"
	"monolith-chi-htmx-sqlite/src/services"
	"monolith-chi-htmx-sqlite/src/validation"

	"github.com/sirupsen/logrus"
)

// Config holds handler configuration
type Config struct {
	DefaultPageSize int
	MaxPageSize     int
}

// OAPIHandler implements the ServerInterface generated by oapi-codegen
type OAPIHandler struct {
	todoService services.TodoServiceInterface
	templates   *template.Template
	config      *Config
	logger      *logrus.Logger
}

// NewOAPIHandler creates a new OAPI handler
func NewOAPIHandler(templates *template.Template, config *Config, todoService services.TodoServiceInterface, logger *logrus.Logger) gen.ServerInterface {
	return &OAPIHandler{
		todoService: todoService,
		templates:   templates,
		config:      config,
		logger:      logger,
	}
}

// GetTodoListPage implements the GetTodoListPage operation
func (h *OAPIHandler) GetTodoListPage(w http.ResponseWriter, r *http.Request, params gen.GetTodoListPageParams) {
	// Extract parameters with defaults
	categoryFilter := ""
	if params.Category != nil {
		categoryFilter = *params.Category
	}

	page := 1
	if params.Page != nil {
		page = *params.Page
	}

	pageSize := h.config.DefaultPageSize
	if params.PageSize != nil {
		pageSize = *params.PageSize
	}

	// Validate pagination parameters
	pageStr := ""
	if params.Page != nil {
		pageStr = strconv.Itoa(*params.Page)
	}
	pageSizeStr := ""
	if params.PageSize != nil {
		pageSizeStr = strconv.Itoa(*params.PageSize)
	}

	_, _, validationResult := validation.ValidatePagination(pageStr, pageSizeStr, h.config.MaxPageSize)
	if !validationResult.IsValid {
		middleware.HandleValidationError(h.logger, w, &middleware.ValidationResult{
			IsValid: false,
			Errors:  validationResult.Errors,
		})
		return
	}

	// Get todos using service
	paginatedTodos, err := h.todoService.GetTodos(categoryFilter, page, pageSize)
	if err != nil {
		middleware.HandleAppError(h.logger, w, err)
		return
	}

	// Get categories
	categories, err := h.todoService.GetCategories()
	if err != nil {
		middleware.HandleAppError(h.logger, w, err)
		return
	}

	data := struct {
		Todos      []models.Todo
		Categories []models.Category
		Filter     string
		Pagination models.Pagination
	}{
		Todos:      paginatedTodos.Todos,
		Categories: categories,
		Filter:     categoryFilter,
		Pagination: paginatedTodos.Pagination,
	}

	if err := h.templates.ExecuteTemplate(w, "index.html", data); err != nil {
		h.logger.WithFields(logrus.Fields{
			"error": err,
		}).Error("Error executing template")
		http.Error(w, "Internal server error", http.StatusInternalServerError)
	}
}

// CreateTodo implements the CreateTodo operation
func (h *OAPIHandler) CreateTodo(w http.ResponseWriter, r *http.Request) {
	if err := r.ParseForm(); err != nil {
		http.Error(w, "Error parsing form", http.StatusBadRequest)
		return
	}

	title := strings.TrimSpace(r.FormValue("title"))
	categoryIDsStr := r.Form["categories"]

	// Convert category IDs to integers
	var categoryIDs []int
	for _, idStr := range categoryIDsStr {
		if id, err := strconv.Atoi(idStr); err == nil && id > 0 {
			categoryIDs = append(categoryIDs, id)
		}
	}

	// Create todo using service
	err := h.todoService.CreateTodo(title, categoryIDs)
	if err != nil {
		middleware.HandleAppError(h.logger, w, err)
		return
	}

	// Redirect to index page
	http.Redirect(w, r, "/", http.StatusSeeOther)
}

// UpdateTodoStatus implements the UpdateTodoStatus operation
func (h *OAPIHandler) UpdateTodoStatus(w http.ResponseWriter, r *http.Request, id int) {
	if err := r.ParseForm(); err != nil {
		http.Error(w, "Error parsing form", http.StatusBadRequest)
		return
	}

	status := r.FormValue("status")

	// Update status using service
	err := h.todoService.UpdateTodoStatus(id, status)
	if err != nil {
		middleware.HandleAppError(h.logger, w, err)
		return
	}

	// Return success response for HTMX
	w.WriteHeader(http.StatusOK)
}

// DeleteTodo implements the DeleteTodo operation
func (h *OAPIHandler) DeleteTodo(w http.ResponseWriter, r *http.Request, id int) {
	// Delete todo using service
	err := h.todoService.DeleteTodo(id)
	if err != nil {
		middleware.HandleAppError(h.logger, w, err)
		return
	}

	// Return success response for HTMX
	w.WriteHeader(http.StatusOK)
}

// CreateCategory implements the CreateCategory operation
func (h *OAPIHandler) CreateCategory(w http.ResponseWriter, r *http.Request) {
	if err := r.ParseForm(); err != nil {
		http.Error(w, "Error parsing form", http.StatusBadRequest)
		return
	}

	title := strings.TrimSpace(r.FormValue("title"))

	// Create category using service
	err := h.todoService.CreateCategory(title)
	if err != nil {
		middleware.HandleAppError(h.logger, w, err)
		return
	}

	// Redirect to index page
	http.Redirect(w, r, "/", http.StatusSeeOther)
}

// DeleteCategory implements the DeleteCategory operation
func (h *OAPIHandler) DeleteCategory(w http.ResponseWriter, r *http.Request, id int) {
	// Delete category using service
	err := h.todoService.DeleteCategory(id)
	if err != nil {
		middleware.HandleAppError(h.logger, w, err)
		return
	}

	// Return success response for HTMX
	w.WriteHeader(http.StatusOK)
} 